pyproject.toml
[tool.poetry]
name = "django-ninja-sample"
version = "0.1.0"
description = ""
authors = ["Iori ONDA <iorionda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
django-ninja = "^1.1.0"
psycopg2 = "^2.9.9"
ruff = "^0.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version="py39"
select = [
    "E",
    "F",
    "W",
    "I",
    "Q",
]
line-length = 119
[tool.ruff.flake8-quotes]
inline-quotes = "single"

[tool.ruff.isort]
force-single-line = true

[tool.ruff.per-file-ignores]
"./tests/*" = ["E101"]